+++ CHANGELOG.md
@@ -39,7 +39,8 @@
 - PR #2850: Clean up unused params in paramsPCA
 - PR #2871: Add timing function to utils
 - PR #2863: in FIL, rename leaf_value_t enums to more descriptive
-- PR #2892 Update ci/local/README.md
+- PR #2892: Update ci/local/README.md
+- PR #2903: Moving linalg's gemm, gemv, transpose to RAFT namespaces
 
 ## Bug Fixes
 - PR #2882: Allow import on machines without GPUs